+++ CHANGELOG.md
@@ -8,9 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - `genomepy search` now accepts the `--exact` flag
+- When installing assemblies from older Ensembl release versions, a clearer error message is given if assembly cannot be found:
+  - if the release does not exist, options will be given
+  - if the assembly does not exist on the release version, all available options are given
+  - if the URL to the genome or annotation files is incorrect, the error message stays the same
 
 ### Fixed
-- `genomepy.install()` now returns a `Genome` instance with updated annotation attributes. 
+- `genomepy.install()` now returns a `Genome` instance with updated annotation attributes.
 
 ## [0.15.0] - 2023-02-28