+++ CHANGELOG.md
@@ -11,9 +11,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - `genomepy.Annotation.attributes()` returns a list of all attributes from the GTF attributes column. 
   - e.g. gene_name, gene_version
   - nice to use with `genomepy.Annotation.from_attributes()` or `genomepy.Annotation.gtf_dict()`
+- When installing assemblies from older Ensembl release versions, a clearer error message is given if assembly cannot be found:
+  - if the release does not exist, options will be given
+  - if the assembly does not exist on the release version, all available options are given
+  - if the URL to the genome or annotation files is incorrect, the error message stays the same
+
+### Changed
+- function `get_division` is now a class method of EnsemblProvider
+- EnsemblProvider class methods `get_division` and `get_version` now require an assembly name.
 
 ### Fixed
-- `genomepy.install()` now returns a `Genome` instance with updated annotation attributes. 
+- `genomepy.install()` now returns a `Genome` instance with updated annotation attributes.
+- now ignoring ~1600 assemblies from the Ensembl database with incorrect metadata
+  - no easy way to retrieve this data
 
 ## [0.15.0] - 2023-02-28