+++ CHANGELOG.md
@@ -15,10 +15,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - if the release does not exist, options will be given
   - if the assembly does not exist on the release version, all available options are given
   - if the URL to the genome or annotation files is incorrect, the error message stays the same
+- new config option: `ucsc_mirror`, options: `eu` or `us`.
+  - the mirror should only affect download speed
+  - can be nice if the other mirror is down!
 
 ### Changed
 - function `get_division` is now a class method of EnsemblProvider
 - EnsemblProvider class methods `get_division` and `get_version` now require an assembly name.
+- UCSC data is now downloaded over HTTPS instead of HTTP
 
 ### Fixed
 - `genomepy.install()` now returns a `Genome` instance with updated annotation attributes.